poc: whole gson tests as separate configs
